[FEAT/#6] 알람뷰, 메인뷰 연결
