+++ capuri-sun/Assets.xcassets/img_alarmBackground.imageset/Contents.json
@@ -5,11 +5,11 @@
       "scale" : "1x"
     },
     {
+      "filename" : "img_alarmBackground.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "(처음)알림주기설정.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
