+++ capuri-sun/View/Main/MainView.swift
@@ -11,12 +11,19 @@ import WeatherKit
 import ActivityKit
 
 struct MainView: View {
-    @StateObject private var locationManager = LocationManager()
-    @State private var location: CLLocation?
-    @State private var address: String = ""
-    @State private var uvIndex: String = ""
-    @State private var condition: String = ""
-    @State private var temperature: String = ""
+    
+//    @StateObject private var locationManager = LocationManager()
+//    @State private var location: CLLocation?
+//    @State private var address: String = ""
+//    @State private var uvIndex: String = ""
+//    @State private var condition: String = ""
+//    @State private var temperature: String = ""
+    
+    @Binding var address: String
+    @Binding var uvIndex: String
+    @Binding var condition: String
+    @Binding var temperature: String
+    
     
     @State private var progress: Double = 0.0
     @State private var timer: Timer?
@@ -28,7 +35,8 @@ struct MainView: View {
     @State private var twoThirdsPassed = false
     
     //@Query var alarms: [Alarm]
-    @AppStorage("alarmTime") var alarmTime: Double = 0.2
+//    @AppStorage("alarmTime") var alarmTime: Double = 0.2
+    @Binding var alarmTime: Double
     
     var body: some View {
         // TODO: 캐릭터 바뀌게 & 시간 변하게
@@ -48,15 +56,15 @@ struct MainView: View {
                             .foregroundColor(.white)
                         
                             .onAppear {
-                                locationManager.getCurrentLocation { location in
-                                    self.location = location
-                                    self.address = locationManager.address
-                                    if let location = location {
-                                        getWeatherInfo(location)
-                                    } else {
-                                        print("위치 정보를 가져올 수 없습니다.")
-                                    }
-                                }
+//                                locationManager.getCurrentLocation { location in
+//                                    self.location = location
+//                                    self.address = locationManager.address
+//                                    if let location = location {
+//                                        getWeatherInfo(location)
+//                                    } else {
+//                                        print("위치 정보를 가져올 수 없습니다.")
+//                                    }
+//                                }
                             }
                         
                         Text(condition)
@@ -424,48 +432,48 @@ struct MainView: View {
         return "\(hours)시간 \(minutes)분"
     }
     
-    func getWeatherInfo(_ location: CLLocation) {
-        Task {
-            do {
-                let service = WeatherService()
-                let result = try await service.weather(for: location)
-                
-                self.uvIndex = "\(result.currentWeather.uvIndex.value)"
-                
-                let temperatureValue = result.currentWeather.temperature.value
-                self.temperature = "\(String(format: "%.1f", temperatureValue))°"
-                
-                self.condition = translateCondition(result.currentWeather.condition.description)
-                
-            } catch {
-                print(String(describing: error))
-            }
-        }
-    }
+//    func getWeatherInfo(_ location: CLLocation) {
+//        Task {
+//            do {
+//                let service = WeatherService()
+//                let result = try await service.weather(for: location)
+//                
+//                self.uvIndex = "\(result.currentWeather.uvIndex.value)"
+//                
+//                let temperatureValue = result.currentWeather.temperature.value
+//                self.temperature = "\(String(format: "%.1f", temperatureValue))°"
+//                
+//                self.condition = translateCondition(result.currentWeather.condition.description)
+//                
+//            } catch {
+//                print(String(describing: error))
+//            }
+//        }
+//    }
     
-    func translateCondition(_ condition: String) -> String {
-        switch condition {
-        case "Partly Cloudy", "Mostly Cloudy", "Cloudy":
-            return "흐림"
-        case "Clear", "Mostly Clear":
-            return "맑음"
-        case "Foggy":
-            return "안개"
-        case "Windy":
-            return "바람"
-        case "Rain", "Heavy Rain", "Drizzle":
-            return "비"
-        case "Snow", "Heavy Snow":
-            return "눈"
-        case "Strong Storms", "Thunder Storms":
-            return "뇌우"
-        default:
-            return condition
-        }
-    }
+//    func translateCondition(_ condition: String) -> String {
+//        switch condition {
+//        case "Partly Cloudy", "Mostly Cloudy", "Cloudy":
+//            return "흐림"
+//        case "Clear", "Mostly Clear":
+//            return "맑음"
+//        case "Foggy":
+//            return "안개"
+//        case "Windy":
+//            return "바람"
+//        case "Rain", "Heavy Rain", "Drizzle":
+//            return "비"
+//        case "Snow", "Heavy Snow":
+//            return "눈"
+//        case "Strong Storms", "Thunder Storms":
+//            return "뇌우"
+//        default:
+//            return condition
+//        }
+//    }
 }
 
-#Preview {
-    MainView()
-}
+//#Preview {
+//    MainView()
+//}
